fix(validatorsecurity): Re-enable EndBlocker with deterministic operations
The EndBlocker was previously disabled to debug consensus issues. After
verifying that all operations are deterministic, the implementation has
been restored.

Determinism verification:
- Uses ctx.BlockTime() (NOT time.Now())
- GetJailedValidators() returns deterministic ordered results
- MonitorAllValidators() iterates in deterministic order
- No map iteration without sorting
- No floating point math
- No crypto/rand usage
- All keeper methods return lexicographically ordered results

The EndBlocker now properly:
1. Runs monitoring checks at configured intervals
2. Logs when jailed validators can unjail themselves (no auto-unjail)

+++ chain/x/monitoring/types/types.go
@@ -125,7 +125,7 @@ type GasPriceTracking struct {
 	MaxPrice        uint64             `json:"max_price"`
 	PriceHistory    []GasPricePoint    `json:"price_history"`
 	TrendDirection  string             `json:"trend_direction"` // increasing, decreasing, stable
-	VolatilityScore float64            `json:"volatility_score"`
+	VolatilityScore uint64             `json:"volatility_score"` // Coefficient of variation in basis points (10000 = 100%)
 }
 
 // GasPricePoint represents a single gas price data point
